Rename CharsetSupportAdditionalIT to CharsetSupportNativeIT and refactor shared logic
