+++ integration-tests/jreleaser-builtin/src/test/java/io/github/bmarwell/jfmt/its/jreleaser/builtin/CharsetSupportIT.java
@@ -6,12 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import io.github.bmarwell.jfmt.its.extension.JFmtTest;
 import io.github.bmarwell.jfmt.its.extension.JdtResult;
-import java.io.IOException;
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 /**
@@ -26,30 +20,7 @@ import org.junit.jupiter.api.Test;
  * 3. Common charsets (UTF-8, ISO-8859-1, Cp1252) are available at runtime
  */
 @JFmtTest
-class CharsetSupportIT {
-
-    private static final Path UTF8_SOURCE_FILE = resolvePath("src/test/resources/charset/Cp1252Test.java");
-    private static final Path CP1252_TEST_FILE = resolvePath("target/test-classes/charset/Cp1252Test.java");
-
-    @BeforeAll
-    static void prepareCp1252TestFile() throws IOException {
-        NativeItConditions.assumeNativeProfileOnGraalVm();
-        Files.createDirectories(CP1252_TEST_FILE.getParent());
-
-        final String source = Files.readString(UTF8_SOURCE_FILE, StandardCharsets.UTF_8);
-        Files.writeString(CP1252_TEST_FILE, source, Charset.forName("windows-1252"));
-    }
-
-    private static Path resolvePath(String moduleRelativePath) {
-        final Path currentDirectory = Path.of("").toAbsolutePath();
-        final Path directPath = currentDirectory.resolve(moduleRelativePath);
-
-        if (Files.exists(directPath)) {
-            return directPath;
-        }
-
-        return currentDirectory.resolve("integration-tests/jreleaser-builtin").resolve(moduleRelativePath);
-    }
+class CharsetSupportIT extends AbstractCharsetIT {
 
     /**
      * Test that the print command can process files with Windows-1252 (Cp1252) characters.
