+++ integration-tests/jreleaser-builtin/src/test/java/io/github/bmarwell/jfmt/its/jreleaser/builtin/CharsetSupportNativeIT.java
@@ -6,39 +6,10 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import io.github.bmarwell.jfmt.its.extension.JFmtTest;
 import io.github.bmarwell.jfmt.its.extension.JdtResult;
-import java.io.IOException;
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 @JFmtTest
-class CharsetSupportAdditionalIT {
-
-    private static final Path UTF8_SOURCE_FILE = resolvePath("src/test/resources/charset/Cp1252Test.java");
-    private static final Path CP1252_TEST_FILE = resolvePath("target/test-classes/charset/Cp1252Test.java");
-
-    @BeforeAll
-    static void prepareCp1252TestFile() throws IOException {
-        NativeItConditions.assumeNativeProfileOnGraalVm();
-        Files.createDirectories(CP1252_TEST_FILE.getParent());
-
-        final String source = Files.readString(UTF8_SOURCE_FILE, StandardCharsets.UTF_8);
-        Files.writeString(CP1252_TEST_FILE, source, Charset.forName("windows-1252"));
-    }
-
-    private static Path resolvePath(String moduleRelativePath) {
-        final Path currentDirectory = Path.of("").toAbsolutePath();
-        final Path directPath = currentDirectory.resolve(moduleRelativePath);
-
-        if (Files.exists(directPath)) {
-            return directPath;
-        }
-
-        return currentDirectory.resolve("integration-tests/jreleaser-builtin").resolve(moduleRelativePath);
-    }
+class CharsetSupportNativeIT extends AbstractCharsetIT {
 
     @Test
     @JFmtTest(args = { "print", "target/test-classes/charset/Cp1252Test.java" })
